+++ src/canns/models/__init__.py
@@ -2,6 +2,7 @@ from .basic import (
     CANN1D as CANN1D,
     CANN1D_SFA as CANN1D_SFA,
     CANN2D as CANN2D,
+    CANN2D_SFA as CANN2D_SFA,
 )
 # from .hybrid import *
 # from .brain_inspired import *
