+++ src/canns/models/basic/__init__.py
@@ -1,8 +1,6 @@
-from .cann1d import (
+from .cann import (
     CANN1D as CANN1D,
     CANN1D_SFA as CANN1D_SFA,
-)
-
-from .cann2d import (
     CANN2D as CANN2D,
+    CANN2D_SFA as CANN2D_SFA,
 )
