+++ src/canns/models/basic/_base.py
@@ -1,4 +1,8 @@
 import brainstate
 
 
-class BasicModel(brainstate.nn.Dynamics): ...
+class BasicModel(brainstate.nn.Dynamics):
+    """
+    Base class for basic models.
+    """
+    pass
