+++ src/canns/task/__init__.py
@@ -1,1 +1,1 @@
-from . import tracking1d as tracking1d
+from . import tracking as tracking
