+++ src/canns/task/_base.py
@@ -2,4 +2,7 @@ from brainstate.nn import Module
 
 
 class BaseTask(Module):
-    ...
+    """
+    Base class for tasks module.
+    """
+    pass
