--- src/canns/task/tracking2d.py
@@ -1,20 +1,0 @@
-from typing import Sequence
-
-import brainstate
-import brainunit as u
-from brainunit import Quantity
-from saiunit.math import dtype
-
-from ._base import BaseTask
-from ..models.basic.cann1d import BaseCANN1D
-
-class PopulationCoding2D():
-    ...
-
-
-class TemplateMatching2D():
-    ...
-
-
-class SmoothTracking2D():
-    ...
